backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
aiosqlite==0.20.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pysnmp==4.4.12